requirements.txt
Babel==1.3
Flask==0.9
Flask-Babel==0.8
Flask-Login==0.2.7
Flask-Mail==0.7.6
Flask-OAuth==0.12
Flask-OpenID==1.1.1
Flask-SQLAlchemy==0.16
Flask-SuperAdmin==1.7.1
Flask-WTF==0.8.4
Flask-WhooshAlchemy==0.54a
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.5
Werkzeug==0.9.4
Whoosh==2.5.4
argparse==1.2.1
blinker==1.3
coverage==3.7
decorator==3.4.0
flup==1.0.2
gdata==2.0.18
httplib2==0.8
oauth2==1.5.211
python-openid==2.2.5
pytz==2013b
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
